Create README.md
To add to this strength, to steer this considerable support in the right direction, we've charted an impressive line-up of well-respected experts drawn from across the globe, as our guiding beacon next. A noteworthy mention among our expert line-up is our friends from the vibrant LGBT community in Poland. The way they've been moulding the movement serves as an indelible mark for the coming years, especially for a transformative 2025, one that promise a horizon aglow with breakthroughs in representation and rights for transgender individuals.